pylinted the code for 8+ score
